REFACTORY fl_controller - Move complete_cycle, average_plan_diffs to cycle_manager - Fix worker.is_eligible method - Update report event
